backend/requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
pydantic
scikit-learn
librosa
hiero-sdk-python
python-multipart
google-generativeai
timm
opencv-python
pillow
facenet-pytorch
numpy<2.0.0
web3
eth-account
python-dotenv
pytesseract
easyocr
mediapipe
redis
onnxruntime
websockets
torchaudio
openai
